+++ webclient/app/components/MotisPaginationControls.vue
@@ -1,0 +1,95 @@
+<script setup lang="ts">
+import { mdiChevronLeft, mdiChevronRight } from "@mdi/js";
+
+interface Props {
+  previousPageCursor?: string | null;
+  nextPageCursor?: string | null;
+  loading?: boolean;
+  size?: 'sm' | 'lg';
+}
+
+const props = withDefaults(defineProps<Props>(), {
+  size: 'sm'
+});
+
+const emit = defineEmits<{
+  loadPrevious: [cursor: string];
+  loadNext: [cursor: string];
+}>();
+
+const { t } = useI18n({ useScope: "local" });
+
+const showPagination = computed(() => {
+  return props.previousPageCursor || props.nextPageCursor;
+});
+
+const handlePreviousPage = () => {
+  if (props.previousPageCursor) {
+    emit("loadPrevious", props.previousPageCursor);
+  }
+};
+
+const handleNextPage = () => {
+  if (props.nextPageCursor) {
+    emit("loadNext", props.nextPageCursor);
+  }
+};
+</script>
+
+<template>
+  <div v-if="showPagination">
+    <!-- Small pagination controls (for top of list) -->
+    <div v-if="size === 'sm'" class="flex items-center gap-2">
+      <button
+        :disabled="!previousPageCursor || loading"
+        @click="handlePreviousPage"
+        class="inline-flex items-center px-2 py-1 text-xs font-medium text-zinc-600 bg-zinc-100 rounded hover:bg-zinc-200 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+      >
+        <MdiIcon :path="mdiChevronLeft" :size="14" class="mr-1" />
+        {{ t("earlier") }}
+      </button>
+      <button
+        :disabled="!nextPageCursor || loading"
+        @click="handleNextPage"
+        class="inline-flex items-center px-2 py-1 text-xs font-medium text-zinc-600 bg-zinc-100 rounded hover:bg-zinc-200 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+      >
+        {{ t("later") }}
+        <MdiIcon :path="mdiChevronRight" :size="14" class="ml-1" />
+      </button>
+    </div>
+
+    <!-- Large pagination controls (for bottom of list) -->
+    <div v-else class="flex justify-center gap-4 mt-6 pt-4 border-t border-zinc-200">
+      <button
+        :disabled="!previousPageCursor || loading"
+        @click="handlePreviousPage"
+        class="inline-flex items-center px-4 py-2 text-sm font-medium text-zinc-700 bg-white border border-zinc-300 rounded-lg hover:bg-zinc-50 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+      >
+        <MdiIcon :path="mdiChevronLeft" :size="16" class="mr-2" />
+        {{ t("load_earlier_connections") }}
+      </button>
+      <button
+        :disabled="!nextPageCursor || loading"
+        @click="handleNextPage"
+        class="inline-flex items-center px-4 py-2 text-sm font-medium text-zinc-700 bg-white border border-zinc-300 rounded-lg hover:bg-zinc-50 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+      >
+        {{ t("load_later_connections") }}
+        <MdiIcon :path="mdiChevronRight" :size="16" class="ml-2" />
+      </button>
+    </div>
+  </div>
+</template>
+
+<i18n lang="yaml">
+de:
+  earlier: Früher
+  later: Später
+  load_earlier_connections: Frühere Verbindungen laden
+  load_later_connections: Spätere Verbindungen laden
+
+en:
+  earlier: Earlier
+  later: Later
+  load_earlier_connections: Load earlier connections
+  load_later_connections: Load later connections
+</i18n>
